requirements.txt
torch
torchvision
pytorch-lightning

# MLOps & Tracking
mlflow
dvc
evidently

# API & UI
fastapi
uvicorn
python-multipart
streamlit

# VLMs
transformers

# Utilities
pyyaml
pandas
scikit-learn
